feat:Image covered with watermark compatible with tilt angle.

+++ src/main/java/org/azir/easywatermark/core/config/WatermarkConfig.java
@@ -36,5 +36,5 @@ public class WatermarkConfig {
     /**
      * An angle, in degrees
      */
-    private double angle = 0;
+    private float angle = 0;
 }
